pyproject.toml
[project]
name = "informer-py"
version = "0.0.3"
description = "chiphell crawler"
authors = [{name = "langchou", email = "your@email.com"}]
dependencies = [
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "pyyaml==6.0.1",
    "loguru==0.7.2",
    "pysocks==1.7.1",
    "SQLAlchemy==2.0.23",
    "aiohttp==3.9.1",
    "python-socks==2.4.3",
    "dingtalk-sdk==1.3.8",
    "fake-useragent==2.1.0",
    "openai==1.72.0",
    "httpx[socks]==0.28.1",
    "flask>=3.1.1",
    "schedule>=1.2.2",
]

[tool.uv]
 